java-refactor-kata/movie-rental-java/src/main/java/movierental/legacy/CustomerLegacy.java
package movierental.legacy;

import java.util.ArrayList;
import java.util.List;

public class CustomerLegacy {

    private String _name;
    private List<RentalLegacy> _rentalLegacies = new ArrayList<RentalLegacy>();

    public CustomerLegacy(String name) {
        _name = name;
    }

    public void addRental(RentalLegacy arg) {
        _rentalLegacies.add(arg);
    }

    public String getName() {
        return _name;
    }

    public String statement() {
        double totalAmount = 0;
        int frequentRenterPoints = 0;
        String result = "Rental Record for " + getName() + "\n";

        for (RentalLegacy each : _rentalLegacies) {
            double thisAmount = 0;

            //determine amounts for each line
            switch (each.getMovie().getPriceCode()) {
                case MovieLegacy.REGULAR:
                    thisAmount += 2;
                    if (each.getDaysRented() > 2)
                        thisAmount += (each.getDaysRented() - 2) * 1.5;
                    break;
                case MovieLegacy.NEW_RELEASE:
                    thisAmount += each.getDaysRented() * 3;
                    break;
                case MovieLegacy.CHILDRENS:
                    thisAmount += 1.5;
                    if (each.getDaysRented() > 3)
                        thisAmount += (each.getDaysRented() - 3) * 1.5;
                    break;
            }

            // add frequent renter points
            frequentRenterPoints++;
            // add bonus for a two day new release rental
            if ((each.getMovie().getPriceCode() == MovieLegacy.NEW_RELEASE) && each.getDaysRented() > 1)
                frequentRenterPoints++;

            // show figures for this rental
            result += "\t" + each.getMovie().getTitle() + "\t" + String.valueOf(thisAmount) + "\n";
            totalAmount += thisAmount;
        }

        // add footer lines
        result += "Amount owed is " + String.valueOf(totalAmount) + "\n";
        result += "You earned " + String.valueOf(frequentRenterPoints) + " frequent renter points";

        return result;
    }
}
